Update arbitrary classes
Updated a number of redundant arbitrary classes with native Tailwind classes. There were a number of inconsistencies that I noticed when using the Tailwind examples as a starting place, so this smoothes things out.

It might be worth considering adding the Radix design system within the Tailwind config files versus using arbitrary values for everything. I found a number of old design system values, but wasn't 100% positive which were used for the primitives when they were first created.